Create README.md
Project Name: Weather Forecast Teller Website
Project Description: This project is a website that provides weather forecasts for a specific location. The user can enter the location in the search bar and the website will display the current weather conditions, as well as a forecast for the next 7 days. The website also includes information about the humidity, wind speed, and chance of precipitation.
Project Goals: The goals of this project are to:
Provide accurate and up-to-date weather forecasts for a specific location.
Make the weather forecast information easy to understand and use.
Create a visually appealing and user-friendly website.
Project Technology: The website is built using HTML, CSS, and JavaScript. The weather forecast data is provided by a third-party API.
Project Status: The project is currently in the development phase. The basic functionality of the website is complete, but there are still some features that need to be added.